requirements.txt
# Client dependencies for Manim MCP
google-generativeai>=0.8.0
mcp>=1.0.0

# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
av==13.1.0
    # via manim
beautifulsoup4==4.13.4
    # via manim
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
click==8.2.0
    # via
    #   cloup
    #   manim
    #   typer
    #   uvicorn
cloup==3.0.7
    # via manim
cython==3.1.0
    # via screeninfo
decorator==5.2.1
    # via manim
fastmcp==1.0
    # via manim-mcp-server (pyproject.toml)
glcontext==3.0.0
    # via moderngl
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   fastmcp
    #   mcp
httpx-sse==0.4.0
    # via mcp
idna==3.10
    # via
    #   anyio
    #   httpx
isosurfaces==0.1.2
    # via manim
manim==0.19.0
    # via manim-mcp-server (pyproject.toml)
manimpango==0.6.0
    # via manim
mapbox-earcut==1.0.3
    # via manim
markdown-it-py==3.0.0
    # via rich
mcp==1.8.0
    # via fastmcp
mdurl==0.1.2
    # via markdown-it-py
moderngl==5.12.0
    # via
    #   manim
    #   moderngl-window
moderngl-window==3.1.1
    # via manim
networkx==3.4.2
    # via manim
numpy==2.2.5
    # via
    #   isosurfaces
    #   manim
    #   mapbox-earcut
    #   moderngl-window
    #   scipy
pillow==11.2.1
    # via
    #   manim
    #   moderngl-window
pycairo==1.28.0
    # via manim
pydantic==2.11.4
    # via
    #   fastmcp
    #   mcp
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.9.1
    # via
    #   fastmcp
    #   mcp
pydub==0.25.1
    # via manim
pyglet==2.1.6
    # via moderngl-window
pyglm==2.8.2
    # via moderngl-window
pygments==2.19.1
    # via
    #   manim
    #   rich
pyobjc-core==11.0
    # via pyobjc-framework-cocoa
pyobjc-framework-cocoa==11.0
    # via screeninfo
python-dotenv==1.1.0
    # via
    #   fastmcp
    #   pydantic-settings
python-multipart==0.0.20
    # via mcp
rich==14.0.0
    # via
    #   manim
    #   typer
scipy==1.15.3
    # via manim
screeninfo==0.8.1
    # via manim
shellingham==1.5.4
    # via typer
skia-pathops==0.8.0.post2
    # via manim
sniffio==1.3.1
    # via anyio
soupsieve==2.7
    # via beautifulsoup4
srt==3.5.3
    # via manim
sse-starlette==2.3.4
    # via mcp
starlette==0.46.2
    # via
    #   mcp
    #   sse-starlette
svgelements==1.9.6
    # via manim
tqdm==4.67.1
    # via manim
typer==0.15.3
    # via fastmcp
typing-extensions==4.13.2
    # via
    #   anyio
    #   beautifulsoup4
    #   manim
    #   pydantic
    #   pydantic-core
    #   typer
    #   typing-inspection
typing-inspection==0.4.0
    # via
    #   pydantic
    #   pydantic-settings
uvicorn==0.34.2
    # via mcp
watchdog==6.0.0
    # via manim
